Added label re-routing
If there is a label before/after the TEXT input control you can click on
it to give focus to the TEXT input.

However this would be lost when the dropdowns [visually] replace the
textbox.

Added enhancement which looks for a related label control and updates
it's FOR attribute to be the first dropdown (usually the day).

Clicking on the label now gives focus to the dropdown.