Added associated time dropdown plug-in
I needed a version that did the same for the time as well as the date.

Follows similar pattern to your date version.  Would seem sensible to
add back into your library in case anyone else finds it useful.

Added a demo file so you can see it in action.